refactor(rome_formatter): Group leading comments
It's necessary to move leading comments outside a group to prevent that they expand the enclosing group:

```javascript
/* a comment */
[1]
```

The `/* a comment */` belongs to the `[` group token that is part of a group wrapping the whole `[1]` expression. It's important that the comment `/* a comment */` gets moved out of the group element to avoid that the `[1]` group expands because of the line break inserted by the comment.

The `group_elements` implementation so far used `split_trivia` to accomplish this.

This PR changes `group_elements` to use a custom implementation that only splits the leading trivia but not the trailing because that isn't necessary for groups to make it clear that the sole use case for `split_trivia` is for cases where the formatter parenthesizes an "expression".